+++ src/RTUB.Application/Services/EnemyState.cs
@@ -9,8 +9,8 @@ internal class EnemyState
 {
     public Character Enemy { get; set; } = null!;
     public int Index { get; set; }
-    public int HP { get; set; }
-    public int MaxHP { get; set; }
+    public long HP { get; set; }
+    public long MaxHP { get; set; }
     public string Name { get; set; } = string.Empty;
     public double ActionTimeMs { get; set; }
     public double Timer { get; set; }
